Breakout integration_test.py
move from a monofile to a more maintainable
testing structure. Also adds retrying to
3dt tests to account for latency
